v0.2.2 eraser not great
psych, I'm not in v0.3 yet, the eraser is not playing well with the fractal frames. I need to make it paint transparent instead of white, but that's a totally different approach so it needs more work. At least I fixed the order of layers issue by changing z orders of strokes vs fractal frames.