v0.3 complete drawing tool (for real)
ok I fixed the opacity issues, and created a brand new eraser system (one eraser line per previous stroke). I have a feeling it's very much not optimized and might slow down when there are a lot of strokes... but for now it works as intended. No idea how I'd optimize with this approach either...